Create a placer module
The placer module will be used to spawn and place the entities in the
game
